Finished testing delegates
